datastructures: allow smaller atomic counters (#201)
This change allows for the use of smaller atomic counters. Rust 1.34 stabilized
fixed width atomics for u8, u16, u32, and u64 which may be more suitable for
specific use-cases than atomic usize. Particularly, for histograms covering wide
ranges of values, but where any given bucket is unlikely to have a large count,
reducing from atomic usize (effectively 64bit atomics on x86_64) to u32 could
yield appreciable savings in memory footprint.

To support this change, we make `Counter` generic across u8, u16, u32, and u64.
Histograms and heatmaps then are able to use these smaller uints to hold bucket
counters. This change also updates dependent crates within the project to make
use of the changes to the datastructures.

+++ rpc-perf/src/client/common.rs
@@ -171,13 +171,13 @@ impl Common {
         }
     }
 
-    pub fn stat_interval(&self, label: Stat, start: usize, stop: usize) {
+    pub fn stat_interval(&self, label: Stat, start: u64, stop: u64) {
         if let Some(ref stats) = self.stats {
             stats.time_interval(label, start, stop);
         }
     }
 
-    pub fn heatmap_increment(&self, start: usize, stop: usize) {
+    pub fn heatmap_increment(&self, start: u64, stop: u64) {
         if let Some(ref stats) = self.stats {
             stats.heatmap_increment(start, stop);
         }
